PAINT_007_001 Implemented Texbox Hex Listener. Moved Code to own class. [MA, PA]
